Add group awareness to command line arguments.
This closes #2 by making the Parameters class aware of the user's
decision to have or not have the group prefixes on their command line
arguments.

This means that command line arguments may be prefixed by group names
(i.e. --foo-bar vs --bar for --bar in group foo).  All tests still pass
including the new tests added to verify this behavior.

The default behavior is to respect groups and use the prefixes but this
can be disabled.