Update all tests to use fixtures.
Now the fixtures have several items as well as use in several tests.
There are still tests that may benefit from fixtures but this simplifies
the majority of cases.  We also added more in-depth tests for addition
of parameters in variuos circumstances.  Including a test of duplicate
arguments being added.