Add definition of ResourceWarning if not present.
Seems that python2 doesn't include ResourceWarning whereas python3 does.
In order to make the code a touch more portable, we're adding that
definition.

Also, this would have been caught if we had already been using travis's
fast fail option to stop as soon as we get a failure.  The flake8 error
indicating this problem would have been more prominent.

I'm getting some weird test output from nose when I run with python2 but
hoping that travis can tell me if it's actually a concern or not.